Move AllCoins components into components react folder && Re-Import at App main component
