+++ src/react/components/AllCoins.jsx
